Removed old actions location and put into store folder
